fix: satisfy participant profile quality gate
